requirements.txt
pyrogram
TgCrypto
pytgcalls
python-dotenv
youtube_dl
youtube_search_python
requests
aiohttp
aiofiles
asyncio
youtube_search
search_engine_parser
ffmpeg
Pillow
python-arq
future
psutil
wget
git+https://github.com/pytgcalls/pytgcalls
